+++ includes/object-cache.php
@@ -1250,7 +1250,7 @@ class WP_Object_Cache {
 
             $results = array_map( function ( $response ) {
                 return (bool) $this->parse_redis_response( $response );
-            }, $tx->{$method}() ?? [] );
+            }, $tx->{$method}() ?: [] );
 
             if ( count( $results ) !== count( $keys ) ) {
                 throw new Exception( 'Redis pipeline returned unexpected result' );
@@ -1499,7 +1499,7 @@ class WP_Object_Cache {
 
             $results = array_map( function ( $response ) {
                 return (bool) $this->parse_redis_response( $response );
-            }, $tx->{$method}() ?? [] );
+            }, $tx->{$method}() ?: [] );
 
             if ( count( $results ) !== count( $keys ) ) {
                 throw new Exception( 'Redis pipeline returned unexpected result' );
@@ -2221,7 +2221,7 @@ LUA;
 
             $results = array_map( function ( $response ) {
                 return (bool) $this->parse_redis_response( $response );
-            }, $tx->{$method}() ?? [] );
+            }, $tx->{$method}() ?: [] );
 
             if ( count( $results ) !== count( $keys ) ) {
                 throw new Exception( 'Redis pipeline returned unexpected result' );